bug dans certains cas de changement d'image OK

+++ backend/controllers/post-controllers.js
@@ -26,34 +26,67 @@ exports.createPost = (req, res, next) => {
 
 exports.modifyPost = (req, res, next) => {
     if (req.file) {
+
         Post.findOne({ where: { id: req.params.id }})
         .then(post => {
+            if (post.image) {
             const filename = post.image.split('/images/')[1];
             fs.unlink(`images/${filename}`, () => {
                 const modifyPost = {
                 title: req.body.title,
                 content: req.body.content,
-                image: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
-            };
+                image: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+                };
     
                 Post.update(modifyPost , { where: { id: req.params.id } })
             
                     .then(() => res.status(200).json({message : 'Post modifié !'}))
                     .catch( error => res.status(400).json({error}));
-        })})
+            })} else {
+                const modifyPost = {
+                title: req.body.title,
+                content: req.body.content,
+                image: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+                };
+        
+                Post.update(modifyPost , { where: { id: req.params.id } })
+        
+                    .then(() => res.status(200).json({message : 'Post modifié !'}))
+                    .catch( error => res.status(400).json({error}));
+            }
+        })
     } else {
-        const modifyPost = {
-        title: req.body.title,
-        content: req.body.content,
-        };
+        Post.findOne({ where: { id: req.params.id }})
+        .then(post => {
+            if (post.image) {
+                const filename = post.image.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    const modifyPost = {
+                    title: req.body.title,
+                    content: req.body.content,
+                    image: ''
+                    };
 
-        Post.update(modifyPost , { where: { id: req.params.id } })
+                    Post.update(modifyPost , { where: { id: req.params.id } })
 
-            .then(() => res.status(200).json({message : 'Post modifié !'}))
-            .catch( error => res.status(400).json({error}));
+                        .then(() => res.status(200).json({message : 'Post modifié !'}))
+                        .catch( error => res.status(400).json({error}));
+                })
+            } else {
+                const modifyPost = {
+                title: req.body.title,
+                content: req.body.content,
+                };
+        
+                Post.update(modifyPost , { where: { id: req.params.id } })
+        
+                    .then(() => res.status(200).json({message : 'Post modifié !'}))
+                    .catch( error => res.status(400).json({error}));
+            }
+        })
     }
-    
-};
+}
+
 
 exports.deletePost = (req, res, next) => {
     Post.findOne({ where: { id: req.params.id }})
